Revert "Make phase equipment checklist editable (#33)"
This reverts commit 3368eeaf317df5dfb776995519a343e005a6bdea.
